backend/requirements.txt
# Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.9.0
pydantic-settings==2.1.0
supabase==2.3.4
python-dotenv==1.0.0
httpx>=0.24,<0.26
sqlalchemy>=2.0.0
pandas-ta>=0.3.14b0

# Core Data/Trading
ccxt
pandas
numpy
